requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai==1.10.0

# URL scraping
cloudscraper==1.2.71
playwright==1.40.0

# HTTP client
httpx==0.26.0

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Image processing and OCR
Pillow==10.2.0
pytesseract==0.3.10

# NLP - spaCy for entity extraction
spacy==3.8.0
# Download model: python -m spacy download en_core_web_sm

# CSV handling with pandas
pandas==2.1.4
openpyxl==3.1.5

# Data validation
pydantic>=2.5.3
pydantic-settings>=2.3.0

# Environment variables
python-dotenv==1.0.1

# Logging
python-json-logger==2.0.7

# CORS
starlette==0.35.1

# Async support
anyio==4.2.0

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development tools (optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# PDF Processing with Docling
docling @ git+https://github.com/IgniteAISolutions/docling-PDF.git@main
